Infer_API/app/requirements.txt
fastapi
uvicorn
numpy
tritonclient[all]
python-multipart
pyngrok
python-dotenv
httpx